A minor change that will be helpful-- make the file generated by the attempt to build a list of classes start with a lower-case character, so that all directories will start with an upper-case character and all files a lower-case character. I've also expanded the list of prefixes that are being removed from the names of the classes.    buildClassList.py: changed output file name to weeklyClassList, added name prefixes    		      to be deleted, cleaned up the code that reduces the name. 		      separateClassFiles.py: changed input file name to weeklyClassList
processClassData.sh: replace the separateFiles.sh and renameInDirectory.sh scripts with
buildClasRoster.py and separateClassFiles.py. The separateClassFiles.py also replaces
the functionality in renameInDirectory.sh, which has been removed from the script. I added
functionality that will call the conversion program to .csv for all of the sql files that
are sorted out, as well.

There are getting to be a lot of junk files in the dumps-- mostly demo or test classes
that have been added for practice or the like. I wrote killListedFiles.py, which currently has
a static list of patterns that, if present in a file name, will cause that file to be
deleted early on. This replaces the shell rm commands; we should make sure that we keep
the list of patterns up to day.